[FIX] Coding style, company form view
