+++ product_internal_reference_code/__manifest__.py
@@ -1,27 +1,15 @@
 # -*- coding: utf-8 -*-
-#    OpenERP, Open Source Management Solution
-#    Copyright (c) Rooms For (Hong Kong) Limited T/A OSCG. All Rights Reserved.
-#
-#    This program is free software: you can redistribute it and/or modify
-#    it under the terms of the GNU Affero General Public License as
-#    published by the Free Software Foundation, either version 3 of the
-#    License, or (at your option) any later version.
-#
-#    This program is distributed in the hope that it will be useful,
-#    but WITHOUT ANY WARRANTY; without even the implied warranty of
-#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#    GNU Affero General Public License for more details.
-#
-#    You should have received a copy of the GNU Affero General Public License
-#    along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
+# Copyright 2017 Quartile Limited
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 {
     'name': 'Product Internal Reference Code',
-    'version': '1.0',
+    'version': '11.0.1.0.0',
     'category': 'Sales',
-    'author': 'Rooms For (Hong Kong) T/A OSCG',
-    'website': 'www.openerp-asia.net',
-    'depends': ['product'],
+    'author': 'Quartile Limited',
+    'website': 'https://www.odoo-asia.com',
+    'depends': [
+        'product',
+    ],
     'description': """ 
     - When product is created, the system should auto-generate a sequential number 
     for default_code
@@ -37,9 +25,9 @@
         default_code based on company configuration.
     """,
     'data': [
-            'data/product_data.xml',
-            'views/company_view.xml',
-             ],
+        'data/product_data.xml',
+        'views/company_view.xml',
+    ],
     'application': False,
     'installable': True,
 }
