+++ product_internal_reference_code/data/product_data.xml
@@ -12,4 +12,3 @@
 
     </data>
 </odoo>
-
