--- product_internal_reference_code/data/product_sequence.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-    <data noupdate="1">
-    
-        <!-- Sequences for Product Internal Code -->
-        <record id="product_internal_code_seq" model = "ir.sequence">
-            <field name="name">Product Sequence</field>
-            <field name="code">product.product.internal_code</field>
-            <field name="padding">5</field>
-            <field eval="False" name="company_id"/>
-        </record>
-        
-    </data>
-</odoo>
-
