+++ product_internal_reference_code/i18n/ja.po
@@ -1,0 +1,31 @@
+# Translation of Odoo Server.
+# This file contains the translation of the following modules:
+#	* product_internal_reference_code
+#
+msgid ""
+msgstr ""
+"Project-Id-Version: Odoo Server 11.0\n"
+"Report-Msgid-Bugs-To: \n"
+"POT-Creation-Date: 2017-11-29 06:38+0000\n"
+"PO-Revision-Date: 2017-11-29 06:38+0000\n"
+"Last-Translator: <>\n"
+"Language-Team: \n"
+"MIME-Version: 1.0\n"
+"Content-Type: text/plain; charset=UTF-8\n"
+"Content-Transfer-Encoding: \n"
+"Plural-Forms: \n"
+
+#. module: product_internal_reference_code
+#: model:ir.model,name:product_internal_reference_code.model_res_company
+msgid "Companies"
+msgstr "会社"
+
+#. module: product_internal_reference_code
+#: model:ir.model.fields,field_description:product_internal_reference_code.field_res_company_internal_code_prefix
+msgid "Internal Reference Prefix"
+msgstr "プロダクト採番方針コード"
+
+#. module: product_internal_reference_code
+#: model:ir.ui.view,arch_db:product_internal_reference_code.view_company_form_inherit_internal_code_prefix
+msgid "Product Internal Code Prefix.. For example A, B, C...."
+msgstr "プロダクトの内部参照採番"
