--- product_internal_reference_code/models/company.py
@@ -1,27 +1,0 @@
-# -*- coding: utf-8 -*-
-# OpenERP, Open Source Management Solution
-# Copyright (c) Rooms For (Hong Kong) Limited T/A OSCG. All Rights Reserved.
-#
-# This program is free software: you can redistribute it and/or modify
-# it under the terms of the GNU Affero General Public License as
-# published by the Free Software Foundation, either version 3 of the
-# License, or (at your option) any later version.
-#
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU Affero General Public License for more details.
-#
-# You should have received a copy of the GNU Affero General Public License
-# along with this program. If not, see <http://www.gnu.org/licenses/>.
-
-from odoo import models, fields
-
-
-class Company(models.Model):
-    _inherit = "res.company"
-
-    internal_code_prefix = fields.Char(
-        string="Internal Reference Prefix",
-        copy=True,
-    )
