+++ product_internal_reference_code/models/res_company.py
@@ -1,0 +1,14 @@
+# -*- coding: utf-8 -*-
+# Copyright 2017 Quartile Limited
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+
+from odoo import models, fields
+
+
+class ResCompany(models.Model):
+    _inherit = "res.company"
+
+    internal_code_prefix = fields.Char(
+        string="Internal Reference Prefix",
+        copy=True,
+    )
