+++ product_internal_reference_code/views/company_view.xml
@@ -1,21 +1,15 @@
 <?xml version="1.0" encoding="utf-8"?>
 <odoo>
-<data>
 
     <record id="view_company_form_inherit_internal_code_prefix" model="ir.ui.view">
         <field name="name">res.company.form.inherit.internal_code_prefix</field>
         <field name="model">res.company</field>
         <field name="inherit_id" ref="base.view_company_form"/>
         <field name="arch" type="xml">
-            <xpath expr="//notebook" position="inside">
-                <page name="prefix_internal_reference" string="Product Prefix">
-                    <group>
-                        <field name="internal_code_prefix" placeholder="Product Internal Code Prefix.. For example A, B, C...."/>
-                    </group>
-                </page>
+            <xpath expr="//field[@name='report_footer']" position="after">
+                <field name="internal_code_prefix" placeholder="Product Internal Code Prefix.. For example A, B, C...."/>
             </xpath>
         </field>
     </record>
 
-</data>
 </odoo>
